fixed join table issues (db.metadata)
