+++ app/models/servers.py
@@ -1,65 +1,65 @@
-# from .db import db, environment, SCHEMA, add_prefix_for_prod
-# from sqlalchemy.ext.declarative import declarative_base
-# from sqlalchemy.schema import Column, ForeignKey, Table
-# from sqlalchemy.orm import relationship
-# from .db import add_prefix_for_prod
+from .db import db, environment, SCHEMA, add_prefix_for_prod
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.schema import Column, ForeignKey, Table
+from sqlalchemy.orm import relationship
+from .db import add_prefix_for_prod
 
 
 
-# Base = declarative_base()
+# Base = declarative_base() <--- USELESS 
 
-# server_users = Table(
-#     'server_users',
-#     Base.metadata,
-#     db.Column('serverId', db.Integer, ForeignKey(add_prefix_for_prod('servers.id')), primary_key=True),
-#     db.Column('userId', db.Integer, ForeignKey(add_prefix_for_prod('users.id')), primary_key=True),
-# )
+server_users = Table(
+    'server_users',
+    db.metadata,
+    db.Column('serverId', db.Integer, db.ForeignKey(add_prefix_for_prod('servers.id')), primary_key=True),
+    db.Column('userId', db.Integer, db.ForeignKey(add_prefix_for_prod('users.id')), primary_key=True),
+)
 
-# server_admins = Table(
-#     'server_admins',
-#     Base.metadata,
-#     db.Column('serverId',db.Integer, ForeignKey(add_prefix_for_prod('servers.id')), primary_key=True),
-#     db.Column('userId', db.Integer, ForeignKey(add_prefix_for_prod('users.id')), primary_key=True),
-# )
+server_admins = Table(
+    'server_admins',
+    db.metadata,
+    db.Column('serverId',db.Integer, db.ForeignKey(add_prefix_for_prod('servers.id')), primary_key=True),
+    db.Column('userId', db.Integer, db.ForeignKey(add_prefix_for_prod('users.id')), primary_key=True),
+)
 
-# class Server(db.Model):
-#     __tablename__ = 'servers'
+class Server(db.Model):
+    __tablename__ = 'servers'
 
-#     if environment == "production":
-#         __table_args__ = {'schema': SCHEMA}
+    if environment == "production":
+        __table_args__ = {'schema': SCHEMA}
 
-#     id = db.Column(db.Integer, primary_key=True)
-#     ownerId = db.Column(db.Integer, nullable=False)
-#     name = db.Column(db.String(255), nullable=False)
-#     preview_img = db.Column(db.String(255))
+    id = db.Column(db.Integer, primary_key=True)
+    ownerId = db.Column(db.Integer, nullable=False)
+    name = db.Column(db.String(255), nullable=False)
+    preview_img = db.Column(db.String(255))
     
-#     users = relationship('User',
-#                         secondary = server_users,
-#                         back_populates = 'servers')
+    users = db.relationship('User',
+                        secondary = server_users,
+                        back_populates = 'servers')
     
-#     admins = relationship('User',
-#                         secondary=server_admins,
-#                         back_populates='servers')
+    admins = db.relationship('User',
+                        secondary=server_admins,
+                        back_populates='servers')
 
-# # class Server_user(db.Model):
-# #     __tablename__ = 'server_users'
+# class Server_user(db.Model):
+#     __tablename__ = 'server_users'
     
-# #     serverId = db.Column(db.Integer, db.ForeignKey('servers.id'))
-# #     userId = db.Column(db.Integer, db.ForeignKey('users.id'))
+#     serverId = db.Column(db.Integer, db.ForeignKey('servers.id'))
+#     userId = db.Column(db.Integer, db.ForeignKey('users.id'))
     
     
     
-# # class Server_admin(db.Model):
-# #     __tablename__ = 'server_admins'
+# class Server_admin(db.Model):
+#     __tablename__ = 'server_admins'
     
-# #     serverId = db.Column(db.Integer, db.ForeignKey('servers.id'))
-# #     userId = db.Column(db.Integer, db.ForeignKey('users.id'))
+#     serverId = db.Column(db.Integer, db.ForeignKey('servers.id'))
+#     userId = db.Column(db.Integer, db.ForeignKey('users.id'))
 
-#     def to_dict(self):
-#         return {
-#             'id': self.id,
-#             'ownerId': self.ownerId,
-#             'name': self.name,
-#             'preview_img': self.preview_img,
+    def to_dict(self):
+        return {
+            'id': self.id,
+            'ownerId': self.ownerId,
+            'name': self.name,
+            'preview_img': self.preview_img,
             
-#         }
+        }
