+++ app/models/user.py
@@ -5,29 +5,12 @@ from flask_login import UserMixin
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.schema import Column, ForeignKey, Table
 from sqlalchemy.types import Integer
+from .servers import server_admins, server_users
 
-Base = declarative_base()
-
-server_users = Table(
-    'server_users',
-    Base.metadata,
-    Column('serverId', Integer, ForeignKey(
-        add_prefix_for_prod('servers.id')), primary_key=True),
-    Column("userId", Integer, ForeignKey(
-        add_prefix_for_prod('users.id')), primary_key=True),
-)
-
-server_admins = Table(
-    'server_admins',
-    Base.metadata,
-    Column('serverId', Integer, ForeignKey(
-        add_prefix_for_prod('servers.id')), primary_key=True),
-    Column('userId', Integer, ForeignKey(
-        add_prefix_for_prod('users.id')), primary_key=True),
-)
+# Base = declarative_base() <--- USELESS 
 
 
-class User(Base, UserMixin):
+class User(db.Model, UserMixin):
     __tablename__ = 'users'
 
     if environment == "production":
@@ -37,12 +20,13 @@ class User(Base, UserMixin):
     username = db.Column(db.String(40), nullable=False, unique=True)
     email = db.Column(db.String(255), nullable=False, unique=True)
     hashed_password = db.Column(db.String(255), nullable=False)
+    
 
-    servers = relationship('Server',
+    servers = db.relationship('Server',
                         secondary=server_users,
                         back_populates='users')
 
-    admin = relationship('Server',
+    admin = db.relationship('Server',
                         secondary=server_admins,
                         back_populates='admins')
 
@@ -67,44 +51,3 @@ class User(Base, UserMixin):
 
 
 
-class Server(Base):
-    __tablename__ = 'servers'
-
-    if environment == "production":
-        __table_args__ = {'schema': SCHEMA}
-
-    id = db.Column(db.Integer, primary_key=True)
-    ownerId = db.Column(db.Integer, nullable=False)
-    name = db.Column(db.String(255), nullable=False)
-    preview_img = db.Column(db.String(255))
-
-    users = relationship('User',
-                        secondary=server_users,
-                    back_populates='servers')
-
-    admins = relationship('User',
-                        secondary=server_admins,
-                        back_populates='admin')
-
-# class Server_user(db.Model):
-#     __tablename__ = 'server_users'
-
-#     serverId = db.Column(db.Integer, db.ForeignKey('servers.id'))
-#     userId = db.Column(db.Integer, db.ForeignKey('users.id'))
-
-
-# class Server_admin(db.Model):
-#     __tablename__ = 'server_admins'
-
-#     serverId = db.Column(db.Integer, db.ForeignKey('servers.id'))
-#     userId = db.Column(db.Integer, db.ForeignKey('users.id'))
-
-
-    def to_dict(self):
-        return {
-            'id': self.id,
-            'ownerId': self.ownerId,
-            'name': self.name,
-            'preview_img': self.preview_img,
-
-        }
